bug: fix hang when last element in stream takes a long time to process
The stream reader assumed there would be a gap between processing the last element in a list & the stream finishing (end event).
This is not always the case - the last element could have an async component or be a very large object which takes a long time
to process.

This PR adds an explicit 'ended' flag to the stream node so the parser knows if the stream has actually ended.

Fixes #8