Dynamic buffer size on binary encode (#10)
The encoder starts with a `chunkSize` (64Kb by default) buffer and add new buffers (chunks) if needed during an encoding. The result is a concat of all chunks as a single buffer. That's allow to reduce memory consumption as well as a good start for further work to make a streaming binary encoding.

Co-authored-by: Roman Dvornov <rdvornov@gmail.com>